Separate the Connectivity and Topology classes
